+++ src/coleslaw.lisp
@@ -36,7 +36,7 @@ Additional args to render CONTENT can be passed via RENDER-ARGS."
   (ensure-directories-exist filepath)
   (with-open-file (out filepath
                    :direction :output
-                   :if-exists :overwrite
+                   :if-exists :supersede
                    :if-does-not-exist :create)
     (write-line page out)))
 
@@ -57,29 +57,18 @@ Additional args to render CONTENT can be passed via RENDER-ARGS."
 (defgeneric deploy (staging)
   (:documentation "Deploy the STAGING dir, updating the .prev and .curr symlinks.")
   (:method (staging)
-    (with-current-directory coleslaw-conf:*basedir*
-      (let* ((dest (deploy *config*))
-             (new-build (rel-path dest "generated/~a" (get-universal-time)))
-             (prev (rel-path dest ".prev"))
-             (curr (rel-path dest ".curr")))
-        (ensure-directories-exist new-build)
-        (run-program "mv ~a ~a" staging new-build)
-        (when (github-pages *config*)
-          (let ((cname-filename (rel-path "" "~a/CNAME" new-build))
-                (stripped-url (puri:uri-host (puri:parse-uri
-                                               (domain *config*)))))
-            (with-open-file (cname cname-filename
-                                   :direction :output
-                                   :if-exists :supersede)
-              (format cname "~a~%" stripped-url))))
-        (when (probe-file prev)
-          (let ((dest (truename prev)))
-            (if (equal prev dest)
-                (delete-file prev)
-                (run-program "rm -R ~a" dest))))
-        (when (probe-file curr)
-          (update-symlink prev (truename curr)))
-        (update-symlink curr new-build)))))
+   ))
+    (let* ((dest (deploy *config*))
+           (new-build (rel-path dest "generated/~a" (get-universal-time)))
+           (prev (rel-path dest ".prev"))
+           (curr (rel-path dest ".curr")))
+      (ensure-directories-exist new-build)
+      (run-program "mv ~a ~a" staging new-build)
+      (when (probe-file prev)
+        (delete-directory-and-files (truename prev) :if-does-not-exist :ignore))
+      (when (probe-file curr)
+        (update-symlink prev (truename curr)))
+      (update-symlink curr new-build))))
 
 (defun main (config-key)
   "Load the user's config section corresponding to CONFIG-KEY, then