+++ src/packages.lisp
@@ -6,6 +6,7 @@
                             #:mappend
                             #:compose)
   (:import-from :closure-template #:compile-template)
+  (:import-from :cl-fad #:delete-directory-and-files)
   (:export #:main
            #:preview
            #:*config*