+++ src/themes.lisp
@@ -22,9 +22,15 @@ any return value other than nil indicates the injection should be added."
     (list :head (injections-for :head)
           :body (injections-for :body))))
 
+(define-condition theme-does-not-exist (error)
+  ((theme :initarg :theme :reader theme))
+  (:report (lambda (c stream)
+             (format stream "Cannot find the theme: '~A'" (theme c)))))
+
 (defun theme-package (name)
-  "Find the package matching the theme NAME."
-  (find-package (string-upcase (concatenate 'string "coleslaw.theme." name))))
+  "Find the package matching the theme NAME or signal THEME-DOES-NOT-EXIST."
+  (or (find-package (string-upcase (concatenate 'string "coleslaw.theme." name)))
+      (error 'theme-does-not-exist :theme name)))
 
 (defun theme-fn (name &optional (package (theme *config*)))
   "Find the symbol NAME inside PACKAGE which defaults to the theme package."